refactor(hal-api): consolidate SharedI2cBus into hal-api
platform-esp32 と platform-rp2040 で重複していた SharedI2cBus を
hal_api::shared_i2c へ移動。両 platform crate は pub use で再エクスポート
するだけとなり、公開 API / 既存の import パスは変わらない。
テストも hal-api 側に 1 箇所に集約。
